+++ src/RuleSet/DeclarationBlock.php
@@ -181,15 +181,13 @@ class DeclarationBlock extends RuleSet
      * Splits shorthand declarations (e.g. `margin` or `font`) into their constituent parts.
      *
      * @return void
+     *
+     * @deprecated This will be removed without substitution in version 10.0.
      */
     public function expandShorthands()
     {
-        // border must be expanded before dimensions
-        $this->expandBorderShorthand();
-        $this->expandDimensionsShorthand();
-        $this->expandFontShorthand();
-        $this->expandBackgroundShorthand();
-        $this->expandListStyleShorthand();
+        trigger_error("Expanding Shorthands will be removed without substitution in version 10.0.");
+        return;
     }
 
     /**