+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Deprecated
 
+- `DeclarationBlock::expandShorthands()` is deprecated and will be removed without substitution in version 10.0.
+
 ### Removed
 
 - Drop support for PHP < 7.2 (#420)